requirements.txt
asgiref==3.6.0
Django==4.1.7
django-better-admin-arrayfield==1.4.2
django-extra-views==0.14.0
django-filter==23.2
django-nested-admin==4.0.2
django-widget-tweaks==1.4.12
Pillow==9.4.0
psycopg2-binary==2.9.5
python-decouple==3.8
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
Unidecode==1.3.6
